requirements.txt
fastapi
sqlalchemy
pydantic
passlib
pyjwt
psycopg2
python-dotenv
razorpay
pytz
